I18N-1480 - Intellum/Evolve integration (#280)
* I18N-1323 - Update Mojito CLI to use OpenAPI spec for rest calls
Generate code from open API specification and reuse the generated models and clients

* I18N-1323 Remove restclient test dependency from webapp module
Removed restclient dependency from webapp module

* Moved test from restclient to webapp

* Removed conflicting dependencies

* I18N-1323 Reuse generated clients and models in restclient module
Reused generated clients and models in restclient module, and updated cli and webapp modules

* Add RepositoryMapper

* 18N-1323 Move generated clients and models to webapp and cli
Moved generated clients and models to webapp and cli

* Remove unnecessary code

* Add profile to generate openapi spec file

* Add maven plugin to check openapi file checksum

* Update documentation

* Moved client and authentication classes from restclient to common module and removed restclient

* Changed plugins messages

* Update documentation

* I18N-1404 Added and modified new classes to synchronize Evolve/Intellum course

* Modified the Evolve service class

* Refactored Evolve's client and service classes

* Completed the Evolve client class

* Refactor EvolveService class

* Updated Evolve Client error message

* Made additional changes to Evolve client

* Moved Evolve classes to webapp module

* Made additional changes to Evolve service class

* Handled earliest updated on date

* Added new unit tests for Evolve service

* Made addtitional changes to Evolve Service test

* Renamed methods in the Evolve Client class

* Added timeout property

* Updated task timeout

* Made changes for the sync job

* Added Sync Job and made UI changes

* Added a scheduled job for Evolve/Intellum synchronization

* Added call to sync Evolve endpoint

* Fixed build issue

* Fixed build issue

* Renamed flyway sql file

* Modified Evolve sync test

* Modified scheduled job manager

* Made changes to Evolve client

* Modified Evolve sync job and configuration classes

* Made changes to Evolve sync job class

* Made Evolve Service class a Spring component

* Removed the logic for retrieving the earliest datetime of untranslated courses

* Fixed Scheduled Job Manager test

* Removed updated on field from course DTO

* Test a single course

* Reverted the change to test a single course

* Removed magic numbers and renamed variables to include time unit

* Made small changes

* Added comment to syncEvolve method

* Passed retry values as arguments to the constructor of the Evolve Client class

* Fixed issue when updating course translations

* Renamed SQL file that adds the Evolve sync scheduled job type

* Renamed conflicting SQL file

* Added course ID to log message

* Added logs

* Fixed the issue with retrieving asset content

* Fixed failing test

* Added retry properties for Evolve sync endpoint

* Adjusted Evolve sync retry default values

+++ common/src/main/java/com/box/l10n/mojito/iterators/ListWithLastPage.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class ListWithLastPage<T> {
   List<T> list;
-  long lastPage;
+  int lastPage;
 
   public List<T> getList() {
     return list;
@@ -14,11 +14,11 @@ public class ListWithLastPage<T> {
     this.list = list;
   }
 
-  public long getLastPage() {
+  public int getLastPage() {
     return lastPage;
   }
 
-  public void setLastPage(long lastPage) {
+  public void setLastPage(int lastPage) {
     this.lastPage = lastPage;
   }
 }
